+++ src/api/routes/content.router.ts
@@ -4,8 +4,8 @@ import { ContentRepository } from '../data/content/content.repository.js';
 import { PostRepository } from '../data/post/post.repository.js';
 import { PostHistoryRepository } from '../data/postHistory/postHistory.repository.js';
 import { projectPrisma } from '../database/prisma/client.js';
-import { asyncHandler } from '../middleware/asyncHandler.js';
-import { authenticatedUser } from '../middleware/auth.js';
+import { asyncHandler } from '../middlewares/asyncHandler.js';
+import { authenticatedUser } from '../middlewares/auth.js';
 import { createContentUseCaseSchema } from '../useCases/content/createContent.schema.js';
 import { CreateContentUseCase } from '../useCases/content/createContent.useCase.js';
 import { updateContentUseCaseSchema } from '../useCases/content/updateContent.schema.js';
