+++ src/api/routes/file.router.ts
@@ -3,9 +3,9 @@ import express, { Request, Response } from 'express';
 import { InvalidPayloadException } from '../../exceptions/invalidPayload.js';
 import { FileRepository } from '../data/file/file.repository.js';
 import { projectPrisma } from '../database/prisma/client.js';
-import { asyncHandler } from '../middleware/asyncHandler.js';
-import { authenticatedUser } from '../middleware/auth.js';
-import { multipartHandler } from '../middleware/multipartHandler.js';
+import { asyncHandler } from '../middlewares/asyncHandler.js';
+import { authenticatedUser } from '../middlewares/auth.js';
+import { multipartHandler } from '../middlewares/multipartHandler.js';
 import { CreateFileUseCase } from '../useCases/file/createFile.useCase.js';
 import { GetFileUseCase } from '../useCases/file/getFile.useCase.js';
 import { getFileUseCaseSchema } from '../useCases/file/getFileUse.schema.js';
