+++ src/api/routes/index.ts
@@ -3,8 +3,8 @@ import express from 'express';
 import { env } from '../../env.js';
 import { expressLogger } from '../../utilities/expressLogger.js';
 import { authConfig } from '../configs/auth.js';
-import { currentSession } from '../middleware/auth.js';
-import { corsMiddleware } from '../middleware/cors.js';
+import { currentSession } from '../middlewares/auth.js';
+import { corsMiddleware } from '../middlewares/cors.js';
 import { content } from './content.router.js';
 import { file } from './file.router.js';
 import { me } from './me.router.js';
