+++ src/api/routes/me.router.ts
@@ -3,8 +3,8 @@ import express, { Request, Response } from 'express';
 import { authConfig } from '../configs/auth.js';
 import { UserRepository } from '../data/user/user.repository.js';
 import { prisma, projectPrisma } from '../database/prisma/client.js';
-import { asyncHandler } from '../middleware/asyncHandler.js';
-import { authenticatedUser } from '../middleware/auth.js';
+import { asyncHandler } from '../middlewares/asyncHandler.js';
+import { authenticatedUser } from '../middlewares/auth.js';
 import { oneWayHash } from '../utilities/oneWayHash.js';
 import { env } from '../../env.js';
 import { MeRepository } from '../data/user/me.repository.js';
