+++ src/api/routes/post.router.ts
@@ -5,8 +5,8 @@ import { ContentRepository } from '../data/content/content.repository.js';
 import { PostRepository } from '../data/post/post.repository.js';
 import { PostHistoryRepository } from '../data/postHistory/postHistory.repository.js';
 import { projectPrisma } from '../database/prisma/client.js';
-import { asyncHandler } from '../middleware/asyncHandler.js';
-import { authenticatedUser } from '../middleware/auth.js';
+import { asyncHandler } from '../middlewares/asyncHandler.js';
+import { authenticatedUser } from '../middlewares/auth.js';
 import { changeStatusUseCaseSchema } from '../useCases/post/changeStatus.schema.js';
 import { ChangeStatusUseCase } from '../useCases/post/changeStatus.useCase.js';
 import { CreatePostUseCase } from '../useCases/post/createPost.useCase.js';
